Agrega carga de mapa al cliente y soluciona pequeños bugs - Agrega metodos necesarios en deserializer, serializer y protocol de common para permitir el envío por socket de los datos del mapa desde el servidor - Agrega funciones necesarias en el connection_handler/client_receiver - Agrega métodos del mundo para crear los nuevos assets - Arregla bug de colisionables en el editor - Los players ahora puede pegarse en el cuerpo (no es necesario el click a la cabeza)

+++ editor/view/asset_selector.h
@@ -14,7 +14,7 @@ class AssetSelector: public QTabWidget {
     Q_OBJECT
 
 public:
-    AssetSelector(const QHash<uint8_t, AssetData>& tiles, const QHash<uint16_t, AssetData>& colliders,
+    AssetSelector(const QHash<uint8_t, AssetData>& tiles, const QHash<uint8_t, AssetData>& colliders,
                   const QHash<uint8_t, AssetData>& npcs, QWidget* parent = nullptr);
 
     ~AssetSelector();
@@ -27,7 +27,7 @@ private:
     Ui::AssetSelector* ui{};
 
     const QHash<uint8_t, AssetData>& tiles;
-    const QHash<uint16_t, AssetData>& colliders;
+    const QHash<uint8_t, AssetData>& colliders;
     const QHash<uint8_t, AssetData>& npcs;
 };
